KNN Imputer for Missing Value
Data preparation is the process of transforming and cleaning raw data to processing. It is an important step to process, involves reformatting data, making corrections to data and the combining of data sets to enrich data.
Missing value imputation is one of the process of data preparation. Here, we use KNN algorithm imputer to fill the missing value.